fix ex4: refactored naming conventions for interfaces, atributes. Changed sleep method from Animal to concrete implementation sinnce all animals sleeps.

+++ src/main/java/org/example/ex4/Mammal.java
@@ -3,7 +3,7 @@ package org.example.ex4;
 public abstract class Mammal extends Animal {
     private final String furColor;
 
-    public Mammal(String name, double age, String furColor) {
+    public Mammal(String name, int age, String furColor) {
         super(name, age);
 
         if (furColor == null || furColor.isBlank()) {
